requirements.txt
pyserial==3.5
customtkinter==5.2.2
pillow==12.1.1